Update README.md
Yesterday, the elevator went out. Heidi called me because it's the only number she has. Who's fault is that? The effort to climb the stairs with a wheelchair and other items is earnest. As a veteran, I'm of service but as a 1st responder, we must have a better plan.